Fix for macosx-version-min issue.
Proposed fix for issue: https://github.com/faithfracture/Apple-Boost-BuildScript/issues/5

I believe the main problem is that the b2 command was not passed an architecture parm (x86). This caused the user-config.jam settings to not be used at all. I debugged into it and none of the settings were being used: compiler path, compiler flags, etc. It was just using the defaults, which happened to work okay.

I made a few other tweaks of passing macosx-version-min into the b2 command, which I don't believe is required since it is also in the user-config.jam file, but it doesn't hurt to explicitly pass it in.

I also changed the toolset to darwin instead of clang. Before it needed to be clang because the darwin settings in user-config.jam weren't being used and using darwin would cause a compile error since the compiler path wasn't set (forcing the toolset to clang would result in a good compiler path). Now that the user-config.jam settings are being used (because architecture is specified for b2), we can be consistent and specify a toolset of darwin for macOS (just like iOs, tvOS, etc.).

I tested that the minimum OS SDK was properly being set via:

otool -l libboost_thread.a  | grep -A 3 LC_VERSION_MIN_MACOSX